Split Login feature from Facebook feature
Add a filter to include FB sdk
WIP on SDK settings
