+++ src/OrchardCore.Modules/OrchardCore.Facebook/Drivers/FacebookSettingsDisplayDriver.cs
@@ -1,36 +1,50 @@
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.Http;
 using OrchardCore.DisplayManagement.Entities;
 using OrchardCore.DisplayManagement.Handlers;
+using OrchardCore.DisplayManagement.Notify;
 using OrchardCore.DisplayManagement.Views;
 using OrchardCore.Environment.Shell;
+using OrchardCore.Facebook.Services;
 using OrchardCore.Facebook.Settings;
 using OrchardCore.Facebook.ViewModels;
 using OrchardCore.Settings;
 
 namespace OrchardCore.Facebook.Drivers
 {
-    public class FacebookLoginSettingsDisplayDriver : SectionDisplayDriver<ISite, FacebookLoginSettings>
+    public class FacebookSettingsDisplayDriver : SectionDisplayDriver<ISite, FacebookSettings>
     {
         private readonly IAuthorizationService _authorizationService;
+        private readonly IDataProtectionProvider _dataProtectionProvider;
         private readonly IHttpContextAccessor _httpContextAccessor;
+        private readonly INotifier _notifier;
+        private readonly IFacebookService _clientService;
         private readonly IShellHost _shellHost;
         private readonly ShellSettings _shellSettings;
 
-        public FacebookLoginSettingsDisplayDriver(
+        public FacebookSettingsDisplayDriver(
             IAuthorizationService authorizationService,
+            IDataProtectionProvider dataProtectionProvider,
+            IFacebookService clientService,
             IHttpContextAccessor httpContextAccessor,
+            INotifier notifier,
             IShellHost shellHost,
             ShellSettings shellSettings)
         {
             _authorizationService = authorizationService;
+            _dataProtectionProvider = dataProtectionProvider;
+            _clientService = clientService;
             _httpContextAccessor = httpContextAccessor;
+            _notifier = notifier;
             _shellHost = shellHost;
             _shellSettings = shellSettings;
         }
 
-        public override async Task<IDisplayResult> EditAsync(FacebookLoginSettings settings, BuildEditorContext context)
+        public override async Task<IDisplayResult> EditAsync(FacebookSettings settings, BuildEditorContext context)
         {
             var user = _httpContextAccessor.HttpContext?.User;
             if (user == null || !await _authorizationService.AuthorizeAsync(user, Permissions.ManageFacebookApp))
@@ -38,28 +52,39 @@ namespace OrchardCore.Facebook.Drivers
                 return null;
             }
 
-            return Initialize<FacebookLoginSettingsViewModel>("FacebookLoginSettings_Edit", model =>
+            return Initialize<FacebookSettingsViewModel>("FacebookSettings_Edit", model =>
             {
-                model.CallbackPath = settings.CallbackPath;
-            }).Location("Content:5").OnGroup(FacebookConstants.Features.Login);
+                var protector = _dataProtectionProvider.CreateProtector(FacebookConstants.Features.Core);
+
+                model.AppId = settings.AppId;
+                if (!string.IsNullOrWhiteSpace(settings.AppSecret))
+                {
+                    model.AppSecret = protector.Unprotect(settings.AppSecret);
+                }
+
+            }).Location("Content:0").OnGroup(FacebookConstants.Features.Core);
         }
 
-        public override async Task<IDisplayResult> UpdateAsync(FacebookLoginSettings settings, BuildEditorContext context)
+        public override async Task<IDisplayResult> UpdateAsync(FacebookSettings settings, BuildEditorContext context)
         {
-            if (context.GroupId == FacebookConstants.Features.Login)
+            if (context.GroupId == FacebookConstants.Features.Core)
             {
                 var user = _httpContextAccessor.HttpContext?.User;
+
                 if (user == null || !await _authorizationService.AuthorizeAsync(user, Permissions.ManageFacebookApp))
                 {
                     return null;
                 }
 
-                var model = new FacebookLoginSettingsViewModel();
+                var model = new FacebookSettingsViewModel();
                 await context.Updater.TryUpdateModelAsync(model, Prefix);
 
                 if (context.Updater.ModelState.IsValid)
                 {
-                    settings.CallbackPath = model.CallbackPath;
+                    var protector = _dataProtectionProvider.CreateProtector(FacebookConstants.Features.Core);
+                    settings.AppId = model.AppId;
+                    settings.AppSecret = protector.Protect(model.AppSecret);
+
                     await _shellHost.ReloadShellContextAsync(_shellSettings);
                 }
             }
