+++ src/OrchardCore.Modules/OrchardCore.Facebook/Services/IFacebookService.cs
@@ -1,0 +1,14 @@
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Threading.Tasks;
+using OrchardCore.Facebook.Settings;
+
+namespace OrchardCore.Facebook.Services
+{
+    public interface IFacebookService
+    {
+        Task<FacebookSettings> GetSettingsAsync();
+        Task UpdateSettingsAsync(FacebookSettings settings);
+        Task<IEnumerable<ValidationResult>> ValidateSettingsAsync(FacebookSettings settings);
+    }
+}
