--- src/OrchardCore.Modules/OrchardCore.Facebook/Settings/FacebookCoreSettings.cs
@@ -1,8 +1,0 @@
-namespace OrchardCore.Facebook.Settings
-{
-    public class FacebookCoreSettings
-    {
-        public string AppId { get; set; }
-        public string AppSecret{ get; set; }
-    }
-}
