+++ src/OrchardCore.Modules/OrchardCore.Facebook/Views/FacebookLoginSettings.Edit.cshtml
@@ -1,4 +1,4 @@
-@using OrchardCore.Facebook.ViewModels;
+@using OrchardCore.Facebook.Login.ViewModels;
 @model FacebookLoginSettingsViewModel;
 
 @*<p class="alert alert-warning">@T["The current tenant will be reloaded when the settings are saved."]</p>*@
